server/pyproject.toml
[project]
name = "mysql-hub-mcp-server"
version = "0.1.0"
description = "MCP Server for MySQL Hub"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pymysql>=1.1.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "groq>=0.4.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "sqlparse>=0.5.3",
    "fastmcp>=2.11.2",
    "tool>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
